+++ src/Generic/Support/FieldFilter.php
@@ -1,0 +1,29 @@
+<?php
+
+/**
+ * @copyright   ©2025 Maatify.dev
+ * @Library     maatify/data-repository
+ * @Project     maatify:data-repository
+ * @author      Mohamed Abdulalim (megyptm) <mohamed@maatify.dev>
+ * @since       2025-11-25
+ * @see         https://www.maatify.dev Maatify.dev
+ * @link        https://github.com/Maatify/data-repository view project on GitHub
+ * @note        Distributed in the hope that it will be useful - WITHOUT WARRANTY.
+ */
+
+declare(strict_types=1);
+
+namespace Maatify\DataRepository\Generic\Support;
+
+/**
+ * Value object representing a single parsed filter condition.
+ */
+final class FieldFilter
+{
+    public function __construct(
+        public string $field,
+        public string $operator,
+        public mixed $value
+    ) {
+    }
+}